Have RoadPoints self clean up when exiting the tree.
If a user deletes an in-use, connected RoadPoint, it will now remove
any connected RoadSegments, and try to clear the references to itself
in any sibling RoadPoints. This now partially makes Delete in the UI
appear functional, although there are still a similar number of error
printouts on the console when this happens, and there may be some in-
fighting with the Network if RoadPoints are added and removed in
realtime in a live scene.