Supporting basic undo and redo for add next/prior point.
History will get lost and yield error printouts on more than one level
of redo (since the nodes are deleted and re-added), but multiple undo's
appear to work fine.

Moving functionality where relevant out of the panel code, and into
the EditorInspectorPlugin to maintain separation. Required some amout of
"signal bubbling up" but this is expected with godot script design.