Created lane tags and initial auto next/prior LaneSegments assignments.
Implemented a tagging system to assign a given lane's next and prior
lanes, which will allow for AI traffic to continue following the lane
even between different Road Segments.

Right now it is hard coded such that a next lane will always be tagged
with the R0 (reverse) or F0 (forward) tag. Next up is to deal with multi
lanes and lane transitions, but the tag assignment already works as
expected once the right tags are in place.