Fixing nearly all start/end LaneSeg points for transitions.
There remains an edge case where you have a lane being removed on one
side at the same time as a lane being added on the other, the lanes get
very oddly shifted. But all other scenarios of shifts (one or multiple)
lanes on one side at a time or on both side (but being added or removed
in the "same direction") works fine.