Created initial connection tools and overlay drawing indicators
Includes necessary utility function to work around the autofix cyclic
functionality that was getting in the way.

Acknowledging there is one bug in this commit whereby clicking on a
RoadSegment to grab a RoadPoint results in not being able to use its
built in 3D transform gizmo as everything is treated as a drag.