Fixing segment generation for flipped RoadPoints connections
If two RoadPoints treat each other as both being their "Next Point" or
both their "Prior Point", we treat the given start or end point as being
flipped so that we can multiply that RP's basis by -1 to flip it around.
This resolves issues of connecting roads that are growing in the
wrong direction.