Run unsnap after translation event completes for proper undo-redo
Similar to the snap operation, this ensures that when users press undo,
the internal engine translation event doesn't put the road back to the
original position but rather where they last left the road before the
unsnapp occurred. This way, the movement back (on undo) to the attached
position happens in the same step that the road container connections
are restored to avoid bad states.

Removed wrong disconnect signal assertion, since there's no validation
on such functions.