Reworked how RoadPoint selection on click works to use gizmo directly
Previously the addon was using a workaround hack of signals and checking
input states in specific moments to see if a click was made on a segment
to then override the selection to the nearest RoadPoint. Now, we are
using the surface tool to construct a gizmo selection widget which is
the road itself that is a child of that RoadPoint, which has the benefit
of always selecting the RoadPoint which will control the decorations for
that segment (but not longer selecting the closest rp), while still
being able to directly click on the handles of the RPs themselves.